make symlink creation idempotent

+++ configs/fedora/create_symlinks.sh
@@ -20,21 +20,20 @@ for dir in "$REPO_CONFIG_DIR"/*; do
         echo "Checking if need to make backup of $targetpath"
         
         # Backup config target
-        if [ -d "$targetpath" ]; then
+        if [ -d "$targetpath" ] && [ ! -L "$targetpath" ]; then
             echo "Making backup of $targetpath"
-            if [ "$DRY_RUN" = false ]; then
+            if [ "$DRY_RUN" = false ] && [ ! -d "$targetpath-backup" ]; then
               mv "$targetpath" "$targetpath-backup"
             fi
         fi
 
         # Create a symbolic link for the directory in the actual .config directory
         echo "Creating symlink for $targetpath"
-        if [ "$DRY_RUN" = false ]; then
+        if [ "$DRY_RUN" = false ] && [ ! -L "$targetpath" -o "$(readlink -f "$targetpath")" != "$(readlink -f "$dir")" ]; then
           ln -sfn "$dir" "$targetpath"
         fi
     fi
 done
-
 
 DOTFILES_DIR="$SCRIPT_DIR/HOME"
 
@@ -45,15 +44,15 @@ for dotfile in "$DOTFILES_DIR"/*.dotfile; do
         filename=".$(basename "$dotfile" .dotfile)"
         echo "Checking if need to make backup of $filename"
 
-        if [ -f "$HOME/$filename" ]; then
+        if [ -f "$HOME/$filename" ] && [ ! -L "$HOME/$filename" ]; then
             echo "Making backup of $filename"
-            if [ "$DRY_RUN" = false ]; then
+            if [ "$DRY_RUN" = false ] && [ ! -f "$HOME/$filename-backup" ]; then
               mv "$HOME/$filename" "$HOME/$filename-backup"
             fi
         fi
 
         echo "Making symlink of $filename"
-        if [ "$DRY_RUN" = false ]; then
+        if [ "$DRY_RUN" = false ] && [ ! -L "$HOME/$filename" -o "$(readlink -f "$HOME/$filename")" != "$(readlink -f "$dotfile")" ]; then
           ln -sfn "$dotfile" "$HOME/$filename"
         fi
     fi
